Criando código ProdutoRepositorio

+++ Atividade1/Models/Produto.cs
@@ -5,7 +5,7 @@
         public int IdProd { get; set; }
         public string Nome { get; set; }
         public string Descricao { get; set; }
-        public decimal Preco { get; set; }
-        public decimal Quantidade { get; set; }
+        public double Preco { get; set; }
+        public int Quantidade { get; set; }
     }
 }
